* sample.c を command/ 以下に配置 	* sample.c の修正
